requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
openai>=1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==9.5.0